refactor: introduced SinchUrlResolversTests and consolidated all resolver tests there, moving them out of the per-service *ConfigurationTests files
