+++ tests/Sinch.Tests/Verification/VerificationConfigurationTests.cs
@@ -9,20 +9,6 @@ namespace Sinch.Tests.Verification
 {
     public class VerificationConfigurationTests
     {
-        [Theory]
-        [InlineData(null, "https://verification.api.sinch.com/")]
-        [InlineData("https://hello.world", "https://hello.world/")]
-        public void ResolveUrl(string urlOverride, string expectedUrl)
-        {
-            var config = new SinchVerificationConfiguration()
-            {
-                AppKey = "key",
-                AppSecret = "secret",
-                UrlOverride = urlOverride,
-            };
-            SinchUrlResolvers.ResolveVerificationUrl(config).ToString().Should().Be(expectedUrl);
-        }
-
         public record VerificationCredentialsMissingTestCaseData(
             string TestName,
             SinchVerificationConfiguration VerificationConfiguration,
